[level 1] Title: 부족한 금액 계산하기, Time: 0.12 ms, Memory: 78 MB -BaekjoonHub

+++ 프로그래머스/1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.java
@@ -1,8 +1,5 @@
 class Solution {
-    public long solution(int price, int money, int count) {
-        // 등차수열의 합 
-        // 첫항: price 
-        // 항의 개수 count 
-        return Math.max((price+(long)price*count) * count / 2-money, 0);
+    public long solution(long price, long money, long count) {
+        return Math.max((price+price*count) * count / 2-money, 0);
     }
 }
